Add files via upload
The selection of longitudinal reinforcement in reinforced concrete beams is traditionally carried out manually, based on trial-and-error procedures. This approach considers commercially available rebars configurations and must comply with design code requirements, such as minimum steel area and bar spacing. Although common in practice, this process is repetitive and prone to errors when evaluating multiple alternatives under constructability and regulatory constraints. The complexity increases when design decisions must simultaneously satisfy normative, practical, and economic criteria simultaneosly must be satisfied. This study presents a practical educational tool for engineering applications, developed in Python, which assists in defining reinforcement distribution during the structural design process. The program accepts user input for the required steel area, beam geometry, maximum allowable bar diameter, and a tolerance margin. Based on this data, it iterates through a matrix of steel areas associated with various commercial bar types and quantities. The algorithm identifies feasible combinations that comply with ACI 318 provisions and prioritizes those with the least excess steel. The tool supports both students and engineers, enabling faster and code-compliant decision-making in reinforced concrete element design.

In traditional methodology, a bar arrangement is initially selected, and then its compliance with minimum spacing requirements is verified. If the configuration does not meet the code, another alternative is tried, and the process is repeated until an acceptable solution is reached. Even when a valid option is found, it must still be verified whether a more efficient combination, such as using two different diameters, exists. This repetitive process largely depends on the experience and judgment of the designer performing the manual calculation. By contrast, the proposed tool automatically discards combinations that do not satisfy user-defined input constraints. It evaluates all viable alternatives, including mixed-diameter solutions, while respecting spacing and minimum area requirements. As a result, it eliminates the iterative trial-and-error process inherent to manual selection. The computational solution improves accuracy, ensures compliance with design codes, and increases overall efficiency in the reinforcement layout process.